CardList Carrousel migrating to realm case uses

+++ app/src/main/java/com/example/daval/cleanrecyclerview/cardSetup/data/local/Realm/entityRealmCardSetup/RCardSetupEntity.kt
@@ -12,4 +12,4 @@ open class RCardSetupEntity(
     var clientName: String = "",
     var clientId: Int = 0,
     @PrimaryKey
-    var id: Int = 0,) : RealmObject()
+    var id: Long = 0,) : RealmObject()
